refactor: removed provider directory also    moved common/api/crypto to pkg/crypto    and moved framework/provider/tinkcrypto to pkg/crypto/tinkcrypto
