+++ pkg/framework/aries/api/protocol.go
@@ -9,7 +9,7 @@ package api
 import (
 	"errors"
 
-	"github.com/hyperledger/aries-framework-go/pkg/common/api/crypto"
+	"github.com/hyperledger/aries-framework-go/pkg/crypto"
 	"github.com/hyperledger/aries-framework-go/pkg/didcomm/common/transport"
 	"github.com/hyperledger/aries-framework-go/pkg/didcomm/dispatcher"
 	vdriapi "github.com/hyperledger/aries-framework-go/pkg/framework/aries/api/vdri"
